// src/shared/types.ts
export type CallType = 'incoming' | 'outgoing' | 'missed' | 'cancelled';

export interface TelegramCallConnection {
  id: string;
  ip: string;
  ipv6?: string;
  port: number;
  peerTag?: string;
  username?: string;
  password?: string;
  isTurn?: boolean;
  isStun?: boolean;
}

export interface CallLog {
  id: number;
  chatId: string;
  peerId: string;
  peerName: string;
  peerUsername?: string;
  peerAvatarUrl?: string;
  type: CallType;
  isVideo: boolean;
  duration: number; // in seconds
  date: number; // unix timestamp
}